Update service discovery behavior of agent role.
Before this change, it could occur, that no services were added during discovery, because no information was yet available. With the updated task, we ensure, that current data is there. Also we only add new services, as a removal of services would indicate a subsequent run of the role and the removal would be undesired there.