Fetch checkmk agent from delegated node
I have the following setup:
- A: Checkmk server
- B: Host that should be configured
- C: Host running Ansible
- D: Host already configured with Ansible

It's possible to delegate the download of the agent to a different host (D).

The tasks `Download * Agent` download the agent from A to D. The next task `Transfer GENERIC or folder-specific * Agent` copies the agent from C to B, which fails if C and D are different hosts.
Therefore I added a fetch task that copies the agent from D to C.